structure mine tab/table files
