+++ services/minespace-web/src/components/dashboard/mine/inspections/Inspections.js
@@ -1,10 +1,12 @@
+// TODO: Remove this when the file is more fully implemented.
 /* eslint-disable */
 
 import React, { Component } from "react";
 import { connect } from "react-redux";
 import { bindActionCreators } from "redux";
 import CustomPropTypes from "@/customPropTypes";
-import { Row, Col, Table, Typography } from "antd";
+import { Row, Col, Typography } from "antd";
+import InspectionsTable from "@/components/dashboard/mine/Inspections/InspectionsTable";
 
 const { Paragraph, Title, Text } = Typography;
 
@@ -14,38 +16,23 @@ const propTypes = {
 
 const defaultProps = {};
 
-const columns = [
-  { title: "Order No.", dataIndex: "order_no", key: "order_no", sorter: true },
-  { title: "Violation", dataIndex: "violation", key: "violation", sorter: true },
-  { title: "Report No.", dataIndex: "report_no", key: "report_no", sorter: true },
-  { title: "Inspection Type", dataIndex: "inspection_type", key: "inspection_type", sorter: true },
-  { title: "Inspector", dataIndex: "inspector", key: "inspector", sorter: true },
-  { title: "Order Status", dataIndex: "order_status", key: "order_status", sorter: true },
-  { title: "Due", dataIndex: "due", key: "due", sorter: true },
-];
-
-const data = [];
-
 export class Inspections extends Component {
+  // TODO: Accurately set isLoaded when file is more properly implemented.
+  state = { isLoaded: true };
+
   render() {
     return (
       <Row>
         <Col>
           <Title level={4}>Inspections</Title>
           <Paragraph>
-            The below table displays all of the{" "}
+            The below table displays all of the&nbsp;
             <Text className="color-primary" strong>
               inspection orders
-            </Text>{" "}
-            associated with this mine.
+            </Text>
+            &nbsp;associated with this mine.
           </Paragraph>
-          <Table
-            size="small"
-            columns={columns}
-            expandedRowRender={(record) => <p style={{ margin: 0 }}>{record.description}</p>}
-            dataSource={data}
-            locale={{ emptyText: "This mine has no inspection data." }}
-          />
+          <InspectionsTable isLoaded={this.state.isLoaded} />
         </Col>
       </Row>
     );
