+++ services/minespace-web/src/components/dashboard/mine/overview/Overview.js
@@ -1,3 +1,4 @@
+// TODO: Remove this when the file is more fully implemented.
 /* eslint-disable */
 
 import React, { Component } from "react";
